Create MahonyUW_AHRS.ino
This is a modified version of the Mahony error approach. 

Instead of using the expected orientation of the Earth's magnetic field vector and Down as reference vectors, it uses West (magnetometer cross accelerometer) and Up, in the standard NWU orientation scheme.  

The filter appears to converge faster than the original Madwick-coded version, presumably because the reference orientations are orthogonal.